Can now compile on Linux
Added a check to see if the end user is compiling from Linux (via the
UNIX variable) which then links with the appropriate libraries.

You'll need X11 if you want to compile this. You may be able to get it
to run on a Raspberry Pi with some tinkering around but I still need to
get that sorted out.